work on unsigned int

+++ includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: ssabbah <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2016/11/11 15:52:58 by ssabbah           #+#    #+#             */
-/*   Updated: 2017/12/24 18:26:08 by ssabbah          ###   ########.fr       */
+/*   Updated: 2017/12/24 19:08:07 by ssabbah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -152,8 +152,8 @@ void				ft_putstr_fd(char const *s, int fd);
 void				ft_putendl_fd(char const *s, int fd);
 void				ft_putnbr_fd(int n, int fd);
 char 				*ft_strrev(char *str);
-void				ft_putnbr_uns(long int n);
+void				ft_putnbr_uns(unsigned long long n);
 char				*ft_strsub(char const *s, unsigned int start, size_t len);
-int					abs_val(int nb);
+long long			abs_val(long long nb);
 
 #endif
