+++ srcs/abs_val.c
@@ -6,13 +6,13 @@
 /*   By: ssabbah <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/12/07 17:31:51 by ssabbah           #+#    #+#             */
-/*   Updated: 2017/12/13 14:42:56 by ssabbah          ###   ########.fr       */
+/*   Updated: 2017/12/24 19:07:39 by ssabbah          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "../includes/libft.h"
 
-int		abs_val(int nb)
+long long		abs_val(long long nb)
 {
 	if (nb < 0)
 		return (-nb);
